Feat : Upload .json or .csv files

+++ src/components/Table.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useRef } from "react";
-import users from "../database/users.data";
 import TableHeadings from "./TableHeadings";
 import Row from "./Row";
 import { useTableData } from "../context/tableData";
@@ -12,8 +11,7 @@ import Paper from "@mui/material/Paper";
 import DeleteButton from "./DeleteButton";
 import { useRowContext } from "../context/RowContext";
 import { useDispatch } from "react-redux";
-import { deleteRow, setTableData } from "../store/slices/table.slice";
-import { v4 as uuidv4 } from "uuid";
+import { deleteRow } from "../store/slices/table.slice";
 
 export default function BasicTable() {
   const dispatch = useDispatch();
@@ -24,21 +22,6 @@ export default function BasicTable() {
   const tableRef = useRef();
 
   useEffect(() => {
-    if (users) {
-      dispatch(setTableData(users));
-    }
-  }, [users]);
-
-  useEffect(() => {
-    const data = JSON.parse(JSON.stringify(users));
-    data?.forEach((data) => {
-      data.unique_key = uuidv4();
-    });
-
-    dispatch(setTableData(data));
-  }, []);
-
-  useEffect(() => {
     if (tableData) {
       setFilteredData(
         tableData.slice(currentPage * rows, currentPage * 10 + rows)
@@ -47,6 +30,12 @@ export default function BasicTable() {
   }, [currentPage, rows, tableData]);
 
   function handleDeleteRow() {
+    const confirm = window.confirm(
+      "Are you sure you want to delete the selected rows?"
+    );
+
+    if (!confirm) return;
+
     dispatch(deleteRow({ rows: rowsToBeDeleted }));
     setRowsToBeDeleted([]);
   }
